Use uploader extractor as base
Required a small change to the configuration schema, since there are now
separate config options in UploaderConfig for queue sizes.
